+++ examples/cargo/experimental/src/main.rs
@@ -12,7 +12,7 @@ use icu::locale::{locale, subtags::region};
 fn main() {
     let names = RegionDisplayNames::try_new(locale!("fr").into(), Default::default())
         .expect("locale 'fr' should be present in compiled data");
-    let name = names.of(region!("DE")).unwrap();
-    assert_eq!(name, "Allemagne");
+    let name = names.of(region!("DE")).expect("DE display name");
+    assert_eq!(name.as_ref(), "Allemagne");
     println!("{name}");
 }